Added ability to TEDS to store variable sized data from a query callback.
So far the query callbacks weren't able to store any additional data on the delayed commands. This prevented operations like creating a new column and assigning it data in the same call. It also caused a performance loss as every multi-row and/or multi-column call was broken down into individual pieces behind the scenes. This changes introduces a scratch buffer that's conceptually alive for a single frame and allows these types of temporary data to be stored as the delayed commands are going to be executed by the end of the frame.

The following notable changes were made:
- Columns can now be added with data. The data is stored in temporary location and copied during the end of a phase, or moved if a move operator callback is provided.
- A thread-safe scratch buffer was added that allows for allocations that are cleared at the end of the frame.
- The scratch buffer is setup to recycle any full blocks once per frame. This includes deleting all stored objects that require doing so.
- The scratch buffer is available for calls that implement the query callback context.
- RemoveRows in the query callback context uses the Mass native "DestroyEntities"-call as TEDS' row handle and Mass' entity handle are an identical size and directly map to eachother.
- AddColumns and RemoveColumns in the query callback context now use the calls to update the composition of an entity instead of multiple calls to update the columns individually. This is a bit more expensive to set up on the callsite, but requires no/less additional memory and is a lot faster to execute at the end of a phase.
- A bug was fixed whereby deleting a row using an invalid row handle would assert.

#jira UE-186821
#rb logan.buchy

[CL 26613676 by ronald koppers in ue5-main branch]

+++ Engine/Plugins/Experimental/TypedElementsDataStorage/Source/TypedElementsDataStorage/Private/TypedElementDatabase.cpp
@@ -110,6 +110,8 @@ void UTypedElementDatabase::OnPreMassTick(float DeltaTime)
 {
 	checkf(IsAvailable(), TEXT("Typed Element Database was ticked while it's not ready."));
 	OnUpdateDelegate.Broadcast();
+	// Recycle any full scratch memory blocks from the previous frame.
+	ScratchBuffer.RecycleBlocks();
 }
 
 TSharedPtr<FMassEntityManager> UTypedElementDatabase::GetActiveMutableEditorEntityManager()
@@ -244,9 +246,10 @@ bool UTypedElementDatabase::BatchAddRow(TypedElementTableHandle Table, TConstArr
 
 void UTypedElementDatabase::RemoveRow(TypedElementRowHandle Row)
 {
-	if (ActiveEditorEntityManager)
+	FMassEntityHandle Entity = FMassEntityHandle::FromNumber(Row);
+	if (ActiveEditorEntityManager && ActiveEditorEntityManager->IsEntityValid(Entity))
 	{
-		if (ActiveEditorEntityManager->IsEntityActive(FMassEntityHandle::FromNumber(Row)))
+		if (ActiveEditorEntityManager->IsEntityBuilt(FMassEntityHandle::FromNumber(Row)))
 		{
 			ActiveEditorEntityManager->DestroyEntity(FMassEntityHandle::FromNumber(Row));
 		}
@@ -612,7 +615,7 @@ void UTypedElementDatabase::UnregisterTickGroup(FName GroupName, EQueryTickPhase
 TypedElementQueryHandle UTypedElementDatabase::RegisterQuery(FQueryDescription&& Query)
 {
 	return (ActiveEditorEntityManager && ActiveEditorPhaseManager)
-		? Queries.RegisterQuery(MoveTemp(Query), *ActiveEditorEntityManager, *ActiveEditorPhaseManager).Handle
+		? Queries.RegisterQuery(MoveTemp(Query), ScratchBuffer, *ActiveEditorEntityManager, *ActiveEditorPhaseManager).Handle
 		: TypedElementInvalidQueryHandle;
 }
 
@@ -726,7 +729,7 @@ void UTypedElementDatabase::PreparePhase(EQueryTickPhase Phase, float DeltaTime)
 {
 	if (ActiveEditorEntityManager)
 	{
-		Queries.RunPhasePreambleQueries(*ActiveEditorEntityManager, Phase, DeltaTime);
+		Queries.RunPhasePreambleQueries(*ActiveEditorEntityManager, ScratchBuffer, Phase, DeltaTime);
 	}
 }
 
@@ -734,7 +737,7 @@ void UTypedElementDatabase::FinalizePhase(EQueryTickPhase Phase, float DeltaTime
 {
 	if (ActiveEditorEntityManager)
 	{
-		Queries.RunPhasePostambleQueries(*ActiveEditorEntityManager, Phase, DeltaTime);
+		Queries.RunPhasePostambleQueries(*ActiveEditorEntityManager, ScratchBuffer, Phase, DeltaTime);
 	}
 }
